Unify project structure
This PR aims to unify the project structure with other projects in company: traits are moved to the `base` subdirectory of corresponding directories.
