+++ src/services/base/mod.rs
@@ -1,0 +1,1 @@
+pub mod validation_service;
