+++ src/services/mod.rs
@@ -1,1 +1,2 @@
-pub mod validation_service;
+pub mod cedar_validation_service;
+pub mod base;
